move invalid checks to hook
